refactor: finish module system rework
